Move VectorField into Plot
