+++ src/display/Plot.tsx
@@ -1,12 +1,15 @@
 import { OfX, OfY } from "./Plot/Simple"
 import { Parametric } from "./Plot/Parametric"
+import { VectorField } from "./Plot/VectorField"
 
 export const Plot = {
   OfX,
   OfY,
   Parametric,
+  VectorField,
 }
 
 import type { OfXProps, OfYProps } from "./Plot/Simple"
 import type { ParametricProps } from "./Plot/Parametric"
-export type { OfXProps, OfYProps, ParametricProps }
+import type { VectorFieldProps } from "./Plot/VectorField"
+export type { OfXProps, OfYProps, ParametricProps, VectorFieldProps }
