+++ src/index.tsx
@@ -9,7 +9,7 @@ export type {
 } from "./origin/CartesianCoordinates"
 
 export { Plot } from "./display/Plot"
-export type { OfXProps, OfYProps, ParametricProps } from "./display/Plot"
+export type { OfXProps, OfYProps, ParametricProps, VectorFieldProps } from "./display/Plot"
 
 export { Line } from "./display/Line"
 export type {
@@ -33,9 +33,6 @@ export type { PointProps } from "./display/Point"
 
 export { Vector } from "./display/Vector"
 export type { VectorProps } from "./display/Vector"
-
-export { VectorField } from "./display/VectorField"
-export type { VectorFieldProps } from "./display/VectorField"
 
 export { Text } from "./display/Text"
 export type { TextProps, CardinalDirection } from "./display/Text"
@@ -64,3 +61,52 @@ export type { Vector2, Matrix } from "./vec"
 export { vec } from "./vec"
 
 export { Transform, type TransformProps } from "./display/Transform"
+
+/**
+ * @deprecated ❌ `VectorField` is now `Plot.VectorField`.
+ *
+ * ```diff
+ * -import {VectorField} from "mafs"
+ * +import {Plot} from "mafs"
+ *
+ * -<VectorField />
+ * +<Plot.VectorField />
+ * ```
+ */
+export const VectorField = () => null
+
+/**
+ * @deprecated ❌ `FunctionGraph` is now `Plot`.
+ *
+ * ```diff
+ * -import {FunctionGraph} from "mafs"
+ * +import {Plot} from "mafs"
+ * ```
+ */
+export const FunctionGraph = {
+  /**
+   * @deprecated ❌ `FunctionGraph.OfX` is now `Plot.OfX`.
+   *
+   * ```diff
+   * -import {FunctionGraph} from "mafs"
+   * +import {Plot} from "mafs"
+   *
+   * -<FunctionGraph.OfX />
+   * +<Plot.OfX />
+   * ```
+   */
+  OfX: null,
+
+  /**
+   * @deprecated ❌ `FunctionGraph.Parametric` is now `Plot.Parametric`.
+   *
+   * ```diff
+   * -import {FunctionGraph} from "mafs"
+   * +import {Plot} from "mafs"
+   *
+   * -<FunctionGraph.Parametric />
+   * +<Plot.Parametric />
+   * ```
+   */
+  Parametric: null,
+}
